Rename iPhone-Index.html to index.html
by default, web servers like Apache look for a file named "index.html" and nothing else. You would have to change this in the web servers config file to point to another file if you wanted that to be the home page. of course, we have to play by GH rules since we are on their platform.